refactor(frontend): move about us page to module
